backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chonkie>=1.2.1",
    "docling>=2.48.0",
    "google-genai>=1.32.0",
    "supabase>=2.18.1",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]
